added helper functions to utils.js
